Move ArrayData to org.apache.spark.sql.catalyst.data.
