+++ sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/util/MapData.scala
@@ -17,6 +17,7 @@
 
 package org.apache.spark.sql.catalyst.util
 
+import org.apache.spark.sql.catalyst.data.ArrayData
 import org.apache.spark.sql.types.DataType
 
 /**
